mpegmap/HTMLSourcingInbandTracks-3.3-1.html

<!doctype html>
<html>
  <head>
    <title>TC 5.1.4-4 EISS TextTrack</title>
    <link rel="author" title="CableLabs" href="html5@cablelabs.com">
    <meta name="flags" content="[requirement flags]">
    <meta name="assert" content="MPEG-2 SPTS, EISS - UA creates TextTrack, attributes set.">
    <meta name="timeout" content="long">
    <script src="testmedia.js"></script>
    <!-- Source HTTP server local copy of World Wide Web Consortium (W3C) Test Harness. -->
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- Script Version: BETA -->
    <!-- About the Media Resource

         To verify DOM elements the details of the test media must be known prior to test execution.
         This test script is bound to the test media file sourced in the script.
         Sourcing a differnt test media file, or modifying that file, may invalidate the test.

         The media is an MPEG-2, Single Program Transport Stream
         containing two EISS streams and one program insertion cue stream.
    -->
  </head>
  <body>
    <p>
      <a href="http://www.w3.org/TR/html5/references.html#refsINBANDTRACKS">Spec
      Reference [INBANDTRACKS]</a>
    </p>
    <video id="v" width="400" height="300" autoplay controls></video>
    <div id="log"></div>
    <script>
      var test = async_test("EISS/EBIF TextTrack attributes.");
      var vid = document.getElementById("v");
      vid.src = mediaServerURL() + "UserPrivateStreams.ts";

      var eissPid = '1768';
      var ebifPid = '1770';
      var kind = 'metadata';
      var language = '';
      var mode = 'disabled';

      var sawTracks = 0;
      vid.textTracks.addEventListener("addtrack", function(event) {
        var track = event.track;

        if (track.id === eissPid) {
          test.step(function() {
            assert_equals(track.id, eissPid, "TextTrackList1.id EISS1");
            assert_equals(track.kind, kind, "TextTrack1.kind");
            assert_equals(track.language, language, "TextTrack1.language");
            assert_equals(track.mode, mode, "TextTrack1.mode");
          }, "EISS TextTrack attributes");
        } else if (track.id === ebifPid) {
          test.step(function() {
            assert_equals(track.id, ebifPid, "TextTrackList2.id EISS2");
            assert_equals(track.kind, kind, "TextTrack2.kind");
            assert_equals(track.language, language, "TextTrack2.language");
            assert_equals(track.mode, mode, "TextTrack2.mode");
          }, "EBIF TextTrack attributes");
        } else {
          return;
        }
        ++sawTracks;
        if (sawTracks >= 2) {
          test.done();
          vid.pause();
        }
      });

      vid.addEventListener("canplaythrough", function() {
        if (!vid.textTracks.getTrackById(eissPid)) {
          test.step(function() {
            assert_unreached("Did not find first EISS TextTrack.id (PID) " + eissPid + " in DOM.");
          }, "Did not find first EISS TextTrack.id (PID) " + eissPid + " in DOM.");
          test.done();
        } else if (!vid.textTracks.getTrackById(ebifPid)) {
          test.step(function() {
            assert_unreached("Did not find second EISS TextTrack.id (PID) " + ebifPid + " in DOM.");
          }, "Did not find second EISS TextTrack.id (PID) " + ebifPid + " in DOM.");
          test.done();
        }
      });
    </script>
  </body>
</html>
